distinguish brushing from brush end
